requirements.txt
appwrite==2.0.0
face-recognition==1.3.0
Pillow==10.0.0
numpy==1.24.3
